Refactor: extract user controls to separate component
