+++ src/modules/dashboard/components/ColonyMembers/MemberControls/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MemberControls';
